chore: rename projectSelector to projectSelectors
